refactor: exporter 제거
