+++ src/components/common/Layout.tsx
@@ -2,7 +2,7 @@ import { PropsWithChildren } from 'react';
 
 import { styled } from 'styled-components';
 
-import { Header } from '../exporter';
+import { Header } from '../Header';
 
 export const Layout = ({ children }: PropsWithChildren) => (
   <Style.Layout>
