--- src/components/exporter.ts
@@ -1,22 +1,0 @@
-export { ShoppingCart } from './pages/ShoppingCart';
-export { Home } from './pages/Home';
-
-export { Header } from './Header';
-
-export { Cart } from './Cart';
-export { CartActions } from './CartActions';
-export { CartItemList } from './CartItemList';
-export { CartItem } from './CartItem';
-export { OrderSummary } from './OrderSummary';
-export { CartSize } from './CartSize';
-
-export { ProductList } from './ProductList';
-export { Product } from './Product';
-
-export { Skeleton } from './common/Skeleton';
-export { Stepper } from './Stepper';
-
-export { Layout } from './common/Layout';
-export { Button } from './common/Button';
-
-export { Checkbox } from './styled';
